src/socialnetwork/requirements.txt
django==1.10
djangcrispy-forms
djangorestframework
